move DictDocumentStore to its own file added a StoreConfig class rename db_collection to collection in mongo document store initialize document store in worker based on the passed store config
